requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
alembic==1.15.2
redis==6.0.0b1
celery==5.5.0
pydantic==2.11.2
pydantic-settings==2.8.1
python-dotenv==1.1.0
pytest==8.3.5
httpx==0.28.1